requirements.txt
opencv-python
Flask
pytest
numpy
google-genai
Pillow
pillow-heif
picamera2
